# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: iHR Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email or employee ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/register-company:
    post:
      operationId: registerCompany
      tags: [auth]
      summary: Register a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRegistrationInput"
      responses:
        "201":
          description: Company registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register-candidate:
    post:
      operationId: registerCandidate
      tags: [auth]
      summary: Register as a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateRegistrationInput"
      responses:
        "201":
          description: Candidate registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies (super admin)
      responses:
        "200":
          description: Companies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Email or Employee ID
        password:
          type: string
    CompanyRegistrationInput:
      type: object
      required: [companyName, email, adminName, password, industry, companySize]
      properties:
        companyName:
          type: string
        email:
          type: string
        adminName:
          type: string
        password:
          type: string
        industry:
          type: string
        companySize:
          type: string
        mobile:
          type: ["string", "null"]
    CandidateRegistrationInput:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        mobile:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        organizations:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/OrganizationSummary"
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: ["string", "null"]
        role:
          type: string
          enum: [candidate, employee, manager, hr, admin, super_admin]
        employeeId:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    OrganizationSummary:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        role:
          type: string
        logoUrl:
          type: ["string", "null"]
    Company:
      type: object
      required: [id, name, slug, industry, status]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        industry:
          type: string
        companySize:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, pending]
        plan:
          type: ["string", "null"]
        employeeCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name, slug, industry]
      properties:
        name:
          type: string
        slug:
          type: string
        industry:
          type: string
        companySize:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
    CompanyUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        companySize:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
...
    Candidate:
      type: object
      required: [id, email, fullName]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        mobile:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        skills:
          type: ["array", "null"]
          items:
            type: string
        experience:
          type: ["integer", "null"]
        linkedinUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        totalApplications:
          type: ["integer", "null"]
        createdAt:
          type: string
        firstName:
          type: ["string", "null"]
        middleName:
          type: ["string", "null"]
        surname:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        maritalStatus:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        personalEmail:
          type: ["string", "null"]
        fathersName:
          type: ["string", "null"]
        mothersName:
          type: ["string", "null"]
        education:
          type: ["object", "null"]
        currentAddress:
          type: ["string", "null"]
        homeAddress:
          type: ["string", "null"]
        emergencyContacts:
          type: ["array", "null"]
          items:
            type: object
        hasReference:
          type: ["string", "null"]
        refereeName:
          type: ["string", "null"]
        refereePhone:
          type: ["string", "null"]
        refereeEmail:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        bankAccountName:
          type: ["string", "null"]
        bankAccountNumber:
          type: ["string", "null"]
        bankIfscCode:
          type: ["string", "null"]
        bankBranchName:
          type: ["string", "null"]
    CandidateUpdate:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: ["array", "null"]
          items:
            type: string
        linkedinUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        middleName:
          type: ["string", "null"]
        surname:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        maritalStatus:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        personalEmail:
          type: ["string", "null"]
        fathersName:
          type: ["string", "null"]
        mothersName:
          type: ["string", "null"]
        education:
          type: ["object", "null"]
        currentAddress:
          type: ["string", "null"]
        homeAddress:
          type: ["string", "null"]
        emergencyContacts:
          type: ["array", "null"]
          items:
            type: object
        hasReference:
          type: ["string", "null"]
        refereeName:
          type: ["string", "null"]
        refereePhone:
          type: ["string", "null"]
        refereeEmail:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        bankAccountName:
          type: ["string", "null"]
        bankAccountNumber:
          type: ["string", "null"]
        bankIfscCode:
          type: ["string", "null"]
        bankBranchName:
          type: ["string", "null"]
...